refactor optimize integration test
