+++ ams/ams-server/src/test/java/com/netease/arctic/ams/server/AmsEnvironment.java
@@ -2,13 +2,16 @@ package com.netease.arctic.ams.server;
 
 import com.netease.arctic.ams.api.CatalogMeta;
 import com.netease.arctic.ams.api.properties.CatalogMetaProperties;
+import com.netease.arctic.ams.api.properties.TableFormat;
 import com.netease.arctic.ams.server.config.ArcticMetaStoreConf;
 import com.netease.arctic.ams.server.service.ServiceContainer;
 import com.netease.arctic.ams.server.util.DerbyTestUtil;
+import com.netease.arctic.catalog.CatalogTestHelpers;
 import com.netease.arctic.optimizer.OptimizerConfig;
 import com.netease.arctic.optimizer.local.LocalOptimizer;
 import com.netease.arctic.table.TableIdentifier;
 import org.apache.commons.io.FileUtils;
+import org.apache.hadoop.conf.Configuration;
 import org.apache.iceberg.relocated.com.google.common.collect.Maps;
 import org.apache.thrift.transport.TTransportException;
 import org.slf4j.Logger;
@@ -150,61 +153,38 @@ public class AmsEnvironment {
     ServiceContainer.getFileInfoCacheService().syncTableFileInfo(tableIdentifier.buildTableIdentifier(), tableType);
   }
 
-  public void createIcebergCatalog(String catalogName, String warehouseDir) {
-    CatalogMeta icebergCatalog = new CatalogMeta();
-    icebergCatalog.setCatalogName(catalogName);
-    icebergCatalog.setCatalogType(CatalogMetaProperties.CATALOG_TYPE_HADOOP);
-    Map<String, String> storageConfigs = Maps.newHashMap();
-    storageConfigs.put("storage.type", "hdfs");
-    storageConfigs.put("hadoop.core.site", "PGNvbmZpZ3VyYXRpb24+PC9jb25maWd1cmF0aW9uPg==");
-    storageConfigs.put("hadoop.hdfs.site", "PGNvbmZpZ3VyYXRpb24+PC9jb25maWd1cmF0aW9uPg==");
-    icebergCatalog.setStorageConfigs(storageConfigs);
-    Map<String, String> authConfigs = Maps.newHashMap();
-    authConfigs.put("auth.type", "simple");
-    authConfigs.put("auth.simple.hadoop_username", "root");
-    icebergCatalog.setAuthConfigs(authConfigs);
+  public void createIcebergHadoopCatalog(String catalogName, String warehouseDir) {
     Map<String, String> properties = Maps.newHashMap();
-    try {
-      Files.createDirectories(Paths.get(warehouseDir));
-    } catch (IOException e) {
-      LOG.error("failed to create iceberg warehouse dir {}", warehouseDir, e);
-      throw new RuntimeException(e);
-    }
+    createDirIfNotExist(warehouseDir);
     properties.put("warehouse", warehouseDir);
-    properties.put("table-formats", "ICEBERG");
-    icebergCatalog.setCatalogProperties(properties);
-    ServiceContainer.getCatalogMetadataService().addCatalog(icebergCatalog);
+    CatalogMeta catalogMeta = CatalogTestHelpers.buildCatalogMeta(catalogName,
+        CatalogMetaProperties.CATALOG_TYPE_HADOOP, properties, TableFormat.ICEBERG);
+    ServiceContainer.getCatalogMetadataService().addCatalog(catalogMeta);
   }
 
-  public void createLocalCatalog(String catalogName, String warehouseDir) {
-    CatalogMeta localCatalog = new CatalogMeta();
-    localCatalog.setCatalogName(catalogName);
-    localCatalog.setCatalogType(CatalogMetaProperties.CATALOG_TYPE_AMS);
-    Map<String, String> storageConfigs = Maps.newHashMap();
-    storageConfigs.put("storage.type", "hdfs");
-    storageConfigs.put("hive.site", "PGNvbmZpZ3VyYXRpb24+PC9jb25maWd1cmF0aW9uPg==");
-    storageConfigs.put("hadoop.core.site", "PGNvbmZpZ3VyYXRpb24+PC9jb25maWd1cmF0aW9uPg==");
-    storageConfigs.put("hadoop.hdfs.site", "PGNvbmZpZ3VyYXRpb24+PC9jb25maWd1cmF0aW9uPg==");
-    localCatalog.setStorageConfigs(storageConfigs);
-    Map<String, String> authConfigs = Maps.newHashMap();
-    authConfigs.put("auth.type", "simple");
-    authConfigs.put("auth.simple.hadoop_username", "root");
-    localCatalog.setAuthConfigs(authConfigs);
+  public void createMixedIcebergCatalog(String catalogName, String warehouseDir) {
     Map<String, String> properties = Maps.newHashMap();
+    createDirIfNotExist(warehouseDir);
+    properties.put("warehouse", warehouseDir);
+    CatalogMeta catalogMeta = CatalogTestHelpers.buildCatalogMeta(catalogName,
+        CatalogMetaProperties.CATALOG_TYPE_AMS, properties, TableFormat.MIXED_ICEBERG);
+    ServiceContainer.getCatalogMetadataService().addCatalog(catalogMeta);
+  }
+
+  public void createMixedHiveCatalog(String catalogName, Configuration hiveConfiguration) {
+    Map<String, String> properties = Maps.newHashMap();
+    CatalogMeta catalogMeta = CatalogTestHelpers.buildHiveCatalogMeta(catalogName,
+        properties, hiveConfiguration, TableFormat.MIXED_HIVE);
+    ServiceContainer.getCatalogMetadataService().addCatalog(catalogMeta);
+  }
+
+  private void createDirIfNotExist(String warehouseDir) {
     try {
       Files.createDirectories(Paths.get(warehouseDir));
     } catch (IOException e) {
       LOG.error("failed to create iceberg warehouse dir {}", warehouseDir, e);
       throw new RuntimeException(e);
     }
-    properties.put("warehouse", warehouseDir);
-    properties.put("table-formats", "MIXED_ICEBERG");
-    localCatalog.setCatalogProperties(properties);
-    ServiceContainer.getCatalogMetadataService().addCatalog(localCatalog);
-  }
-  
-  public void createCatalog() {
-    
   }
 
   private void stopAms() {
